Work around for 'OSError: [Errno 2] No such file or directory' thrown by 'shutil'

+++ mlt/utils/git_helpers.py
@@ -18,10 +18,12 @@
 # SPDX-License-Identifier: EPL-2.0
 #
 
-from contextlib import contextmanager
-from mlt.utils import process_helpers
+import os
 import shutil
 import tempfile
+from contextlib import contextmanager
+
+from mlt.utils import process_helpers
 
 
 @contextmanager
@@ -33,4 +35,5 @@ def clone_repo(repo):
     try:
         yield destination
     finally:
-        shutil.rmtree(destination)
+        if os.path.exists(destination):
+            shutil.rmtree(destination)
